+++ AmplifyPlugins/Core/AWSPluginsCore/Model/Support/SelectionSet.swift
@@ -38,7 +38,7 @@ extension SelectionSet {
         fields.forEach { field in
             if field.isEmbeddedType, let embeddedTypeSchema = field.embeddedTypeSchema {
                 let child = SelectionSet(value: .init(name: field.name, fieldType: .embedded))
-                child.withCodableFields(embeddedTypeSchema.sortedFields)
+                child.withEmbeddableFields(embeddedTypeSchema.sortedFields)
                 self.addChild(settingParentOf: child)
             } else if field.isAssociationOwner,
                 let associatedModelName = field.associatedModelName,
@@ -54,11 +54,11 @@ extension SelectionSet {
         addChild(settingParentOf: .init(value: .init(name: "__typename", fieldType: .value)))
     }
 
-    func withCodableFields(_ fields: [ModelField]) {
+    func withEmbeddableFields(_ fields: [ModelField]) {
         fields.forEach { field in
             if field.isEmbeddedType, let embeddedTypeSchema = field.embeddedTypeSchema {
                 let child = SelectionSet(value: .init(name: field.name, fieldType: .embedded))
-                child.withCodableFields(embeddedTypeSchema.sortedFields)
+                child.withEmbeddableFields(embeddedTypeSchema.sortedFields)
                 self.addChild(settingParentOf: child)
             } else {
                 self.addChild(settingParentOf: .init(value: .init(name: field.name, fieldType: .value)))
